+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema[8.1].define(version: 2026_06_19_165009) do
+ActiveRecord::Schema[8.1].define(version: 2026_06_19_201954) do
   create_table "action_text_rich_texts", force: :cascade do |t|
     t.text "body"
     t.datetime "created_at", null: false
@@ -188,6 +188,7 @@ ActiveRecord::Schema[8.1].define(version: 2026_06_19_165009) do
     t.integer "created_by_id", null: false
     t.text "description"
     t.datetime "discarded_at"
+    t.json "enabled_tools", default: [], null: false
     t.string "name", null: false
     t.string "slug", null: false
     t.datetime "updated_at", null: false